+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Prevent storing NULL in char and text fields (#274)
 * Update to the latest release of Salmon
+* Make sure user creation and username change forms use the same validation (#281)
 
 ## Releases